% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRMT.R
\name{adaptive.threshold.calipers}
\alias{adaptive.threshold.calipers}
\title{Obtain adaptive thresholds with caliper c specified for each outcome}
\usage{
adaptive.threshold.calipers(
  pair.data,
  num.outcome,
  calipers,
  clinical.threshold = NULL
)
}
\arguments{
\item{pair.data}{A data frame for pairwise comparison,
returned by \code{data.to.pair()}.}

\item{num.outcome}{A numeric value of the number of outcomes.}

\item{clinical.threshold}{A numeric vector represents the minimal clinical thresholds.
Default to 0 for all endpoints.}

\item{caliper}{A numeric vector indicates the caliper c used for each outcome.}
}
\value{
A list of outputs:
\describe{
\item{threshold}{A numeric vector of obtained adaptive thresholds.}
\item{stage}{A numeric vector indicating outcome for each comparison stage.}
}
}
\description{
This function calculates adaptive thresholds for WRAT analysis with
caliper c specified for each outcome. To be called by function \code{WRAT()}.
}
