% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRMT.R
\name{comparison.result}
\alias{comparison.result}
\title{Summarize comparison results for outputs}
\usage{
comparison.result(comparison.output)
}
\arguments{
\item{comparison.output}{A list returned by \code{WRMT.comparison()} or \code{WRAT.comparison()}.
Must contain elements \code{pair.data}, \code{stage.results},
\code{treatment}, and \code{strata}.}
}
\value{
A list of analysis results:
\describe{
\item{WR}{A numeric value. Estimated win ratio for treatment versus control.}
\item{P}{A numeric value. Two-sided P value of the Finkelstein-Schoenfeld (FS) test.}
\item{Stage.decompose}{A list of stage-level comparison results for each stratum.
Each stratum corresponds to a matrix. A single matrix will be returned for
non-stratified analysis for simplicity. The decomposition matrix presents
proportions of wins, ties, and losses (obtained by the treatment group) at each stage.
Proportions are always obtained with respect to the total number of comparisons between
treatment and control groups.}
\item{strata.summary}{A data frame presents strata weights and stratum-specific
WR results.}
\item{FS.stat}{A numeric value of the FS test statistic.}
\item{FS.stat.var}{A numeric value of the variance of the FS test statistic.}
}
}
\description{
This function analyzes pairwise comparisons results provided by
\code{WRMT.comparison()} or \code{WRAT.comparison()}. To be called
by function \code{WRMT()} or \code{WRAT()}.
}
\examples{
# Example 1: stratified comparison
strata.compare <- WRMT.comparison(data = sample_data[,1:6],
                                  threshold = c(0,0), stage = c(1,2),
                                  direction = c(1,1),
                                  strata = sample_data$strata,
                                  strata.weight = NULL)
analysis.result <- comparison.result(strata.compare)
cat("WR:",analysis.result$WR, "; P-value:", analysis.result$P,"\n")

# Example 2: non-stratified comparison
nonstrata.compare <- WRMT.comparison(data = sample_data[,1:6],
                                     threshold = c(0,0), stage = c(1,2),
                                     direction = c(1,1),
                                     strata = NULL,
                                     strata.weight = NULL)
analysis.result <- comparison.result(nonstrata.compare)
cat("WR:",analysis.result$WR, "; P-value:", analysis.result$P,"\n")

}
