% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRMT.R
\name{data.to.pair}
\alias{data.to.pair}
\title{Organize dataset for pairwise comparison}
\usage{
data.to.pair(data)
}
\arguments{
\item{data}{Input data frame}
}
\value{
A data frame for pairwise comparison
}
\description{
Organize dataset for pairwise comparison
}
