% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WRAT}
\alias{WRAT}
\title{WRAT Analysis}
\usage{
WRAT(
  data,
  method = c("weights", "calipers"),
  weights = NULL,
  calipers = NULL,
  clinical.threshold = NULL,
  direction = NULL,
  strata = NULL,
  strata.weight = NULL
)
}
\arguments{
\item{data}{A data frame with each row representing a participant.
The columns should be orders as:
\itemize{
\item \code{ID}: Participant ID, must be numeric and unique.
\item \code{Treatment Group Indicator}: 1 (True) = Treatment, 0 (False) = Control.
\item \code{Outcome 1}: Observed outcome 1 in numeric format.
\item \code{Censor Status of Outcome 1}: 1 (True) = Censored, 0 (False) = Not Censored.
\item \code{Outcome 2}: Observed outcome 2 in numeric format.
\item \code{Censor Status of Outcome 2}: 1 (True) = Censored, 0 (False) = Not Censored.
\item similar for outcome 3, 4, and more.
}}

\item{method}{Method used to obtain adaptive thresholds.
"weights" or "calipers" as detailed below.}

\item{weights}{A numeric vector indicates the weights when method "weights" is employed.
Default to 1 for all outcomes.}

\item{calipers}{A numeric value (or vector) indicates the calipers.

For \code{method="weights"}, a single value that will be applied to all outcomes.
Default to 0.2.

For \code{method="calipers"}, a vector that contains calipers for each outcome.
Input required for \code{method="calipers"}.}

\item{clinical.threshold}{A numeric vector represents the minimal clinical thresholds.
Default to 0 for all endpoints.}

\item{direction}{A numeric vector the more favorable results of each outcome.
1 for favoring larger outcome values (e.g., surviving longer time) and -1 for favoring
smaller outcome values (e.g., having fewer adverse events).}

\item{strata}{A numeric vector indicates the stratum that each participant belongs to.
Only needed for analysis with stratification. Default to leaving blank for non-stratified
analysis.}

\item{strata.weight}{A data frame indicates weights of each stratum.
Default to weighting base on numbers of participants within strata.
If specified, the columns should be orders as:
\itemize{
\item \code{stratum}: Stratum index, should be numeric,
unique and corresponds to \code{strata}.
\item \code{weight}: Weight for each stratum, should be numeric.
}}
}
\value{
A list of analysis results:
\describe{
\item{WR}{A numeric value. Estimated win ratio for treatment versus control.}
\item{P}{A numeric value. Two-sided P value of the Finkelstein-Schoenfeld (FS) test.}
\item{Stage.decompose}{A list of stage-level comparison results for each stratum.
Each stratum corresponds to a matrix. A single matrix will be returned for
non-stratified analysis for simplicity. The decomposition matrix presents
proportions of wins, ties, and losses (obtained by the treatment group) at each stage.
Proportions are always obtained with respect to the total number of comparisons between
treatment and control groups.}
\item{strata.summary}{A data frame presents strata weights and stratum-specific
WR results.}
\item{FS.stat}{A numeric value of the FS test statistic.}
\item{FS.stat.var}{A numeric value of the variance of the FS test statistic.}
}
}
\description{
Perform WRAT (win ratio with adaptive thresholds) analysis.
}
\examples{
# Example 1: stratified WRAT analysis with two time-to-event endpoints.
#   The "weights" method employed.
analysis.results <- WRAT(data = sample_data[,1:6],
                         method = "weights",
                         weights = c(1,1), calipers = 0.2,
                         clinical.threshold = c(0,0),
                         direction = c(1,1),
                         strata = sample_data$strata,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")

# Example 2: stratified WRAT analysis with two time-to-event endpoints.
#   The "calipers" method employed.
analysis.results <- WRAT(data = sample_data[,1:6],
                         method = "calipers",
                         weights = NULL, calipers = c(0.2,0.3),
                         clinical.threshold = c(0,0),
                         direction = c(1,1),
                         strata = sample_data$strata,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")

# Example 3: non-stratified WRAT analysis with two time-to-event endpoints.
#   The "weights" method employed.
analysis.results <- WRAT(data = sample_data[,1:6],
                         method = "weights",
                         weights = c(1,1), calipers = 0.2,
                         clinical.threshold = c(0,0),
                         direction = c(1,1),
                         strata = NULL,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")
}
