% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRMT.R
\name{WRMT.one.stage.sur}
\alias{WRMT.one.stage.sur}
\title{Comparison for a single stage with potential right censoring}
\usage{
WRMT.one.stage.sur(t1, t2, c1, c2, threshold)
}
\arguments{
\item{t1}{A numeric vector. Observed outcome for target participants.}

\item{t2}{A numeric vector. Observed outcome for participants for comparison.}

\item{c1}{A numeric vector. Censoring indicator for target participants.}

\item{c2}{A numeric vector. Censoring indicator for participants for comparison.}

\item{threshold}{A numeric value. Threshold used for comparison at this stage,
differences smaller than this threshold are considered undetermined.}
}
\value{
A numeric vector that indicates the comparison results for this stage:
1 for wins by target participants, -1 for losses by target participants,
and 0 for undetermined pairs.
}
\description{
Comparison for a single stage with potential right censoring
}
