% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{WRMT}
\alias{WRMT}
\title{WRMT Analysis}
\usage{
WRMT(
  data,
  threshold,
  stage = NULL,
  direction = NULL,
  strata = NULL,
  strata.weight = NULL
)
}
\arguments{
\item{data}{A data frame with each row representing a participant.
The columns should be orders as:
\itemize{
\item \code{ID}: Participant ID, must be numeric and unique.
\item \code{Treatment Group Indicator}: 1 (True) = Treatment, 0 (False) = Control.
\item \code{Outcome 1}: Observed outcome 1 in numeric format.
\item \code{Censor Status of Outcome 1}: 1 (True) = Censored, 0 (False) = Not Censored.
\item \code{Outcome 2}: Observed outcome 2 in numeric format.
\item \code{Censor Status of Outcome 2}: 1 (True) = Censored, 0 (False) = Not Censored.
\item similar for outcome 3, 4, and more.
}}

\item{threshold}{A numeric vector indicating thresholds for each stage.}

\item{stage}{A numeric vector indicating the outcome measure considered at
each stage of comparison (1 for outcome 1, 2 for outcome 2, etc.).
If specified, the length must be the same with the length of threshold parameter.
Default to alternating across outcomes when the length of threshold is a multiply of
number of outcomes. E.g., with two outcomes and four thresholds specified,
the default stage is (1,2,1,2).}

\item{direction}{A numeric vector the more favorable results of each outcome.
1 for favoring larger outcome values (e.g., surviving longer time) and -1 for favoring
smaller outcome values (e.g., having fewer adverse events).
efault to favoring larger values for all outcomes.}

\item{strata}{A numeric vector indicates the stratum that each participant belongs to.
Only needed for analysis with stratification. Default to leaving blank for non-stratified
analysis.}

\item{strata.weight}{A data frame indicates weights of each stratum.
Default to weighting base on numbers of participants within strata.
If specified, the columns should be orders as:
\itemize{
\item \code{stratum}: Stratum index, should be numeric,
unique and corresponds to \code{strata}.
\item \code{weight}: Weight for each stratum, should be numeric.
}}
}
\value{
A list of analysis results:
\describe{
\item{WR}{A numeric value. Estimated win ratio for treatment versus control.}
\item{P}{A numeric value. Two-sided P value of the Finkelstein-Schoenfeld (FS) test.}
\item{Stage.decompose}{A list of stage-level comparison results for each stratum.
Each stratum corresponds to a matrix. A single matrix will be returned for
non-stratified analysis for simplicity. The decomposition matrix presents
proportions of wins, ties, and losses (obtained by the treatment group) at each stage.
Proportions are always obtained with respect to the total number of comparisons between
treatment and control groups.}
\item{strata.summary}{A data frame presents strata weights and stratum-specific
WR results.}
\item{FS.stat}{A numeric value of the FS test statistic.}
\item{FS.stat.var}{A numeric value of the variance of the FS test statistic.}
}
}
\description{
Perform WRMT (win ratio with multiple thresholds) analysis.
}
\examples{
# Example 1: stratified WRMT analysis with two time-to-event endpoints.
#   Stages: endpoint 1 with threshold 10, endpoint 2 with threshold 7,
#   endpoint 1 with thresholds 0, and endpoint 2 with threshold 0.
#   That is stratified WRMT(10,7,0,0) analysis.

analysis.results <- WRMT(data = sample_data[,1:6],
                         threshold = c(0,0), stage = c(1,2),
                         direction = c(1,1),
                         strata = NULL,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")


# Example 2: non-stratified WRMT analysis with two time-to-event endpoints.
#   Stages: endpoint 1 with threshold 10, endpoint 2 with threshold 7,
#   endpoint 1 with thresholds 0, and endpoint 2 with threshold 0.
#   That is non-stratified WRMT(10,7,0,0) analysis.

analysis.results <- WRMT(data = sample_data[,1:6],
                         threshold = c(0,0), stage = c(1,2),
                         direction = c(1,1),
                         strata = NULL,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")

# Example 3: stratified standard Win Ratio analysis with
#   two time-to-event endpoints
analysis.results <- WRMT(data = sample_data[,1:6],
                         threshold = c(0,0), stage = c(1,2),
                         direction = c(1,1),
                         strata = sample_data$strata,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")

# Example 4: non-stratified standard Win Ratio analysis with
#   two time-to-event endpoints
analysis.results <- WRMT(data = sample_data[,1:6],
                         threshold = c(0,0), stage = c(1,2),
                         direction = c(1,1),
                         strata = NULL,
                         strata.weight = NULL)
cat("WR:",analysis.results$WR, "; P-value:", analysis.results$P,"\n")

}
